+++ stroom-util/src/main/java/stroom/util/task/taskqueue/TaskExecutor.java
@@ -24,14 +24,13 @@ import stroom.util.spring.StroomShutdown;
 import stroom.util.thread.CustomThreadFactory;
 import stroom.util.thread.StroomThreadGroup;
 
+import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ConcurrentSkipListSet;
 import java.util.concurrent.Executor;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicReference;
 import java.util.concurrent.locks.Condition;
 import java.util.concurrent.locks.ReentrantLock;
 
@@ -44,8 +43,7 @@ public abstract class TaskExecutor {
 
     private final AtomicInteger totalThreads = new AtomicInteger();
 
-    private final ConcurrentSkipListSet<TaskProducer> producers = new ConcurrentSkipListSet<>();
-    private final AtomicReference<TaskProducer> lastProducer = new AtomicReference<>();
+    private final RoundRobinCollection<TaskProducer> producers = new RoundRobinCollection<>();
 
     private final ReentrantLock taskLock = new ReentrantLock();
     private final Condition condition = taskLock.newCondition();
@@ -108,7 +106,7 @@ public abstract class TaskExecutor {
     private String getDebugMessage() {
         final StringBuilder sb = new StringBuilder("Timeout waiting for:");
 
-        for (final TaskProducer producer : producers) {
+        for (final TaskProducer producer : producers.list()) {
             if (producer != null) {
                 sb.append("\n\t");
                 sb.append(producer.toString());
@@ -162,13 +160,8 @@ public abstract class TaskExecutor {
         try {
             if (total < maxThreads) {
                 // Try and get a task from usable producers.
-                final int tries = producers.size();
-                for (int i = 0; i < tries && task == null; i++) {
-                    final TaskProducer producer = nextProducer();
-                    if (producer != null) {
-                        task = producer.next();
-                    }
-                }
+                final Optional<TaskProducer> optionalTaskProducer = producers.next();
+                task = optionalTaskProducer.map(TaskProducer::next).orElse(null);
 
                 final Runnable currentTask = task;
                 if (currentTask != null) {
@@ -195,30 +188,6 @@ public abstract class TaskExecutor {
         }
 
         return task;
-    }
-
-    private TaskProducer nextProducer() {
-        TaskProducer current;
-        TaskProducer producer;
-
-        do {
-            current = lastProducer.get();
-            producer = current;
-            try {
-                if (producer == null) {
-                    producer = producers.first();
-                } else {
-                    producer = producers.higher(producer);
-                    if (producer == null) {
-                        producer = producers.first();
-                    }
-                }
-            } catch (final Exception e) {
-                // Ignore.
-            }
-        } while (!lastProducer.compareAndSet(current, producer));
-
-        return producer;
     }
 
     public void setMaxThreads(final int maxThreads) {